+++ packages/components/organisms/f-footer/CHANGELOG.md
@@ -4,6 +4,17 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 
+v4.8.1
+------------------------------
+*January 29, 2021*
+### Changed
+- Fixed a condition to show country selector
+
+### Added
+- Tests to cover this change
+- Styles for right aligned footer row
+
+
 v4.8.0
 ------------------------------
 *January 28, 2021*